requirements.txt
# Audio processing
librosa
torchaudio
soundfile
numpy
scipy
matplotlib
pandas

# Deep learning
torch
torchvision

# Optional: Transformers (for Transformer ASR experiments)
transformers

# Hyperparameter tuning
optuna
hyperopt
keras-tuner

# Utilities
tqdm
python-Levenshtein

# Optional audio augmentations
audiomentations
torchcodec